+++ integ/target_exceptions.json
@@ -1,7 +1,10 @@
 {
-  "baseline": 2295,
-  "baseline_reason": "Every dropped target counted here predates the gate and each needs its own decision, so --strict fails on a rise rather than demanding zero. Lower this number whenever a case file regains a backend; the gate fails on a drop too, so an improvement cannot be silently spent. The unit is one (case, target) pair, never (file, target): a sequence-style file states one scenario across prep/act/verify cases, so a backend dropped from only the verifying case still runs every step that cannot fail and silently skips the assertion, and a union over the file would report that backend as covered. What this gate exists to catch is the failure mode that hid two real bugs: a backend that cannot pass a case is normally just deleted from that case's `targets`, which turns a divergence into an omission no reviewer can see. integ/unix/mv/empty_dir.json had dropbox and dropbox-root removed for a folder-conflict divergence, and s3/gridfs removed because a directory rename was unimplemented on both -- neither omission said so anywhere. A large share of the remaining count is the tier2/unicode/escape family, whose cases pin text-processing behavior that cannot vary by backend and so run on the core four; those are excusable in bulk once someone confirms that file by file, which is exactly the decision this gate is meant to force rather than assume.",
-  "files": {},
+  "baseline": 2289,
+  "baseline_reason": "Every dropped target counted here predates the gate and each needs its own decision, so --strict fails on a rise rather than demanding zero. Lower this number whenever a case file regains a backend; the gate fails on a drop too, so an improvement cannot be silently spent. The unit is one (case, target) pair, never (file, target): a sequence-style file states one scenario across prep/act/verify cases, so a backend dropped from only the verifying case still runs every step that cannot fail and silently skips the assertion, and a union over the file would report that backend as covered. What this gate exists to catch is the failure mode that hid two real bugs: a backend that cannot pass a case is normally just deleted from that case's `targets`, which turns a divergence into an omission no reviewer can see. integ/unix/mv/empty_dir.json had dropbox and dropbox-root removed for a folder-conflict divergence, and s3/gridfs removed because a directory rename was unimplemented on both -- neither omission said so anywhere. A large share of the remaining count is the tier2/unicode/escape family, whose cases pin text-processing behavior that cannot vary by backend and so run on the core four; those are excusable in bulk once someone confirms that file by file, which is exactly the decision this gate is meant to force rather than assume. Lowered 2295 -> 2289 when cmp/{n,b,i}.json landed: the gate compares each case against its directory's MODAL target set, and cmp's mode was decided by a 2-vs-2 tie broken toward the larger set, so the two fixture-reading cases (basic.json, which only reads /data/a.txt) set the expectation and the two write-seeding ones (l.json, s.json) each reported a 3-target omission. The new cases seed the same way l.json does -- `mkdir -p` plus `printf >` -- which databricks, databricks-prefix and sharepoint-prefix cannot serve, so they carry l.json's list and the write-seeding shape is now the majority. Those 6 pairs are a real capability limit rather than a parked divergence: cmp itself runs on all 22 (basic.json proves it), only the seeding does not. Nothing was narrowed to make a case pass.",
+  "files": {
+    "integ/unix/gzip/f.json": "Same list as gzip/k.json, the sibling that pins -k. -f only decides whether an existing archive is overwritten, which every backend in that list already exercises through the write primitive.",
+    "integ/unix/ls/t.json": "-t sorts on mtime, so the case has to set three distinct mtimes with touch -d first; the list is touch/d.json's, the backends that support setting one."
+  },
   "entries": {
     "integ/bash/assign/export_attr.json :: cd_exports_pwd_and_oldpwd": {
       "hf": "hf holds no empty directory: the service refuses a directory marker client-side (create_dir=false, probed on opendal 0.47.1), so mkdir is a no-op and a directory exists exactly while it holds a key. This case has to cd into a directory it just made and never keyed, so on hf the cd fails and PWD/OLDPWD never take the values being asserted. What the case pins is the export attribute on the two variables cd writes, which is shell state and cannot vary by backend; hf runs the rest of this file, including the bare `declare -p PWD` case that reads the same attribute without needing a directory.",
