+++ integ/unix/checksum/tag.json
@@ -1,0 +1,108 @@
+{
+  "cases": [
+    {
+      "id": "checksum_tag_names_md5",
+      "seq": 910042,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "mkdir -p /data/qcktag && printf 'hello' > /data/qcktag/h.txt && md5sum --tag /data/qcktag/h.txt",
+      "expect": {
+        "exit": 0,
+        "stdout": "MD5 (/data/qcktag/h.txt) = 5d41402abc4b2a76b9719d911017c592\n",
+        "stderr": ""
+      },
+      "flags": [
+        "tag"
+      ]
+    },
+    {
+      "id": "checksum_tag_names_sha1",
+      "seq": 910043,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "sha1sum --tag /data/qcktag/h.txt",
+      "expect": {
+        "exit": 0,
+        "stdout": "SHA1 (/data/qcktag/h.txt) = aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d\n",
+        "stderr": ""
+      },
+      "flags": [
+        "tag"
+      ]
+    },
+    {
+      "id": "checksum_tag_names_sha384",
+      "seq": 910044,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "sha384sum --tag /data/qcktag/h.txt",
+      "expect": {
+        "exit": 0,
+        "stdout": "SHA384 (/data/qcktag/h.txt) = 59e1748777448c69de6b800d7a33bbfb9ff1b463e44354c3553bcdb9c666fa90125a3c79f90397bdf5f6a13de828684f\n",
+        "stderr": ""
+      },
+      "flags": [
+        "tag"
+      ]
+    },
+    {
+      "id": "checksum_tag_names_sha512",
+      "seq": 910045,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "sha512sum --tag /data/qcktag/h.txt",
+      "expect": {
+        "exit": 0,
+        "stdout": "SHA512 (/data/qcktag/h.txt) = 9b71d224bd62f3785d96d46ad3ea3d73319bfbc2890caadae2dff72519673ca72323c3d99ba5c11d7c7acc6e14b8c5da0c4663475c2e5c3adef46f73bcdec043\n",
+        "stderr": ""
+      },
+      "flags": [
+        "tag"
+      ]
+    },
+    {
+      "id": "checksum_short_binary_and_text_markers",
+      "seq": 910046,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "sha256sum -b /data/qcktag/h.txt; sha256sum -t /data/qcktag/h.txt",
+      "expect": {
+        "exit": 0,
+        "stdout": "2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824 */data/qcktag/h.txt\n2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824  /data/qcktag/h.txt\n",
+        "stderr": ""
+      },
+      "flags": [
+        "b",
+        "t"
+      ]
+    },
+    {
+      "id": "checksum_tag_output_checks_back",
+      "seq": 910047,
+      "targets": [
+        "ram",
+        "disk"
+      ],
+      "command": "mkdir -p /data/qckrt && sha512sum --tag /data/qcktag/h.txt > /data/qckrt/sums && sha512sum -c /data/qckrt/sums; echo rc=$?",
+      "expect": {
+        "exit": 0,
+        "stdout": "/data/qcktag/h.txt: OK\nrc=0\n",
+        "stderr": ""
+      },
+      "flags": [
+        "tag",
+        "c"
+      ]
+    }
+  ]
+}
