+++ integ/unix/gzip/f.json
@@ -1,0 +1,38 @@
+{
+  "cases": [
+    {
+      "id": "gzip_f_overwrites_an_existing_archive",
+      "seq": 601307,
+      "targets": [
+        "ram",
+        "disk",
+        "redis",
+        "opfs",
+        "s3",
+        "gridfs",
+        "s3-prefix",
+        "gridfs-prefix",
+        "hf",
+        "hf-prefix",
+        "dropbox",
+        "dropbox-root",
+        "onedrive",
+        "sharepoint",
+        "ssh",
+        "nextcloud",
+        "gdrive",
+        "gdrive-folder",
+        "box"
+      ],
+      "command": "mkdir -p /data/qgzf && printf 'first' > /data/qgzf/g.txt && gzip -k /data/qgzf/g.txt && printf 'second' > /data/qgzf/g.txt && gzip -f /data/qgzf/g.txt && gzip -dc /data/qgzf/g.txt.gz",
+      "expect": {
+        "exit": 0,
+        "stdout": "second",
+        "stderr": ""
+      },
+      "flags": [
+        "f"
+      ]
+    }
+  ]
+}
