+++ integ/unix/ls/t.json
@@ -1,0 +1,41 @@
+{
+  "cases": [
+    {
+      "id": "ls_t_sorts_newest_first",
+      "seq": 960096,
+      "targets": [
+        "ram",
+        "disk",
+        "redis"
+      ],
+      "command": "mkdir -p /data/qlst && printf 'x' > /data/qlst/old.txt && printf 'x' > /data/qlst/mid.txt && printf 'x' > /data/qlst/new.txt && touch -d '2021-01-01T00:00:00' /data/qlst/old.txt && touch -d '2022-01-01T00:00:00' /data/qlst/mid.txt && touch -d '2023-01-01T00:00:00' /data/qlst/new.txt && ls -1t /data/qlst",
+      "expect": {
+        "exit": 0,
+        "stdout": "new.txt\nmid.txt\nold.txt\n",
+        "stderr": ""
+      },
+      "flags": [
+        "t"
+      ]
+    },
+    {
+      "id": "ls_rt_reverses_the_mtime_sort",
+      "seq": 960097,
+      "targets": [
+        "ram",
+        "disk",
+        "redis"
+      ],
+      "command": "ls -1rt /data/qlst",
+      "expect": {
+        "exit": 0,
+        "stdout": "old.txt\nmid.txt\nnew.txt\n",
+        "stderr": ""
+      },
+      "flags": [
+        "r",
+        "t"
+      ]
+    }
+  ]
+}
