+++ python/mirage/commands/builtin/generic/crossmount/relay/cmp.py
@@ -15,10 +15,10 @@
 import functools
 
 from mirage.commands.builtin.generic.cmp import cmp_cmd as generic_cmp
+from mirage.commands.builtin.generic.cmp import parse_flags
 from mirage.commands.builtin.generic.crossmount.types import CrossResult
 from mirage.commands.builtin.generic.crossmount.utils import flat_scopes, relay
-from mirage.commands.spec import SPECS
-from mirage.commands.spec.types import FlagValue, FlagView
+from mirage.commands.spec.types import FlagValue
 from mirage.runtime.types import DispatchFn
 from mirage.types import PathSpec
 
@@ -27,21 +27,22 @@ async def run_cmp(scopes: list[PathSpec], flag_kwargs: dict[str, FlagValue],
                   dispatch: DispatchFn) -> CrossResult:
     """Byte-compare two files on different mounts via the shared generic.
 
-    Pure wiring: both sides are read through dispatch-relayed primitives.
+    Pure wiring: both sides are read through dispatch-relayed primitives,
+    and the flags go through the generic's own ``parse_flags`` -- reading
+    them a second time here is how the relay came to take ``-i`` as a
+    bare int while the generic had moved on to GNU's ``SKIP1:SKIP2``.
 
     Args:
         scopes (list[PathSpec]): The two path operands.
         flag_kwargs (dict): Flags parsed against the shared cmp spec.
         dispatch (DispatchFn): Workspace operation dispatcher.
     """
-    fl = FlagView(flag_kwargs, spec=SPECS["cmp"])
-    limit = fl.as_str("n")
-    skip = fl.as_str("i")
+    parsed = parse_flags(flag_kwargs)
     return await generic_cmp(flat_scopes(scopes),
                              read_bytes=functools.partial(
                                  relay, dispatch, "read"),
-                             silent=fl.as_bool("s"),
-                             verbose=fl.as_bool("args_l"),
-                             limit=int(limit) if limit is not None else None,
-                             print_bytes=fl.as_bool("b"),
-                             skip=int(skip) if skip is not None else None)
+                             silent=parsed.silent,
+                             verbose=parsed.verbose,
+                             limit=parsed.limit,
+                             print_bytes=parsed.print_bytes,
+                             skip=parsed.skip)
