+++ typescript/packages/core/src/workspace/ungated_attr_writes.test.ts
@@ -42,7 +42,7 @@ const ALLOWED: Record<string, string> = {
   'workspace/executor/builtins/vars.ts::handleReadonly':
     'the `=` branch only; `await view.set(key, val)` runs first and ' +
     'the bare form uses `view.mark`',
-  'workspace/node/execute_node.ts::stampExport':
+  'workspace/node/declaration.ts::stampExport':
     'the `covered` branch only, which is the names that carried a ' +
     'value or a staged array literal; a bare name has no gated write ' +
     'to ride on and goes through `view.mark`',
